Move truncate's unit tests into its file
